requirements.txt
allure-pytest == 2.13.2
pytest == 7.4.0
faker == 19.3.1
requests == 2.31.0
jsonschema~=4.18.4
selenium==4.11.0